refactor(product): repository layer refactor by interface
