+++ src/main/java/gift/product/service/ProductService.java
@@ -3,22 +3,20 @@ package gift.product.service;
 import gift.product.domain.Product;
 import gift.product.dto.ProductRequestDto;
 import gift.product.exception.ProductNotFoundException;
+import gift.product.repository.MemoryProductRepository;
 import gift.product.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicLong;
 
 @Service
 @Transactional
 public class ProductService {
     private final ProductRepository repository;
 
-    public ProductService(ProductRepository repository){
+    public ProductService(MemoryProductRepository repository){
         this.repository = repository;
     }
